hw2

+++ app/src/main/java/com/example/cinema/model/Cinema.kt
@@ -3,5 +3,5 @@ package com.example.cinema.model
 data class Cinema(
     val movie: String,
     val released: Int,
-    val rating: Int
+    val rating: Double
     )
